requirements.txt
# Core dependencies
agent-framework-devui
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Configuration and data handling
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI/ML dependencies
openai>=1.0.0
azure-openai>=1.0.0
anthropic>=0.7.0

# Database and storage (optional)
sqlalchemy>=2.0.0
alembic>=1.12.0

# Monitoring and logging
structlog>=23.0.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Azure specific
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0

# Additional utilities
click>=8.1.0
rich>=13.0.0
httpx>=0.25.0
